Update ExpansionPanel.js
Can we use componentWillReceiveProps for dynamically setting/updating the isControlled property? This would be useful for me. I have a list of ExpansionPanel's that I'm dynamically generating inside of a Dialog component, and whenever I close the Dialog box, I want all of the ExpansionPanel's to be closed automatically, so when I reopen the Dialog box, all Panel's are closed. Unless there is a solution to this that I'm missing. Thanks!